Improve dir structure;
