+++ internal/api/handlers/courses/httpmodels/review.go
@@ -1,0 +1,52 @@
+package httpmodels
+
+import (
+	"cybersafe-backend-api/internal/models"
+	"net/http"
+	"time"
+
+	"github.com/asaskevich/govalidator"
+	"github.com/google/uuid"
+	"gorm.io/gorm"
+)
+
+type ReviewFields struct {
+	Comment string `json:"comment" valid:"required"`
+	Rating  int    `json:"rating" valid:"range(1|5), required"`
+
+	CourseID uuid.UUID `json:"courseID" valid:"required"`
+}
+
+type ReviewResponse struct {
+	ReviewFields
+
+	ID        uuid.UUID      `json:"id" valid:"uuid, required"`
+	CreatedAt time.Time      `json:"createdAt"`
+	UpdatedAt time.Time      `json:"updatedAt"`
+	DeletedAt gorm.DeletedAt `json:"deletedAt"`
+
+	UserID uuid.UUID `json:"userID"`
+}
+
+type ReviewRequestContent struct {
+	ReviewFields
+}
+
+func (rre *ReviewRequestContent) Bind(_ *http.Request) error {
+	_, err := govalidator.ValidateStruct(*rre)
+
+	if err != nil {
+		return err
+	}
+
+	return err
+}
+
+func (rrc *ReviewRequestContent) ToEntityReview() *models.Review {
+	review := &models.Review{
+		Comment:  rrc.Comment,
+		Rating:   rrc.Rating,
+		CourseID: rrc.CourseID,
+	}
+	return review
+}
